fix(chat): preserve compatibility authority builds
