+++ internal/corecmd/corecmd_test.go
@@ -2224,3 +2224,80 @@ func TestCrossPlatformCoverageEmbedContractSkipsBlankAndHiddenFlags(t *testing.T
 		}
 	}
 }
+
+func TestCrossPlatformCoverageReviewedHiddenAliasTarget(t *testing.T) {
+	newCommand := func() *cobra.Command {
+		cmd := &cobra.Command{Use: "run"}
+		cmd.Flags().String("public", "", "public")
+		cmd.Flags().String("alias", "", "alias")
+		_ = cmd.Flags().MarkHidden("alias")
+		return cmd
+	}
+	annotate := func(cmd *cobra.Command, target, origin string) {
+		flag := cmd.Flags().Lookup("alias")
+		if target != "" {
+			runtimeannotate.SetFlagAnnotation(flag, runtimeannotate.AnnotationFlagAliasOf, target)
+		}
+		if origin != "" {
+			runtimeannotate.SetFlagAnnotation(flag, runtimeannotate.AnnotationFlagAliasOrigin, origin)
+		}
+	}
+
+	if _, ok, err := ReviewedHiddenAliasTarget(nil, "alias"); ok || err != nil {
+		t.Fatalf("nil alias = ok %v, err %v", ok, err)
+	}
+	cmd := newCommand()
+	if _, ok, err := ReviewedHiddenAliasTarget(cmd, "public"); ok || err != nil {
+		t.Fatalf("public flag = ok %v, err %v", ok, err)
+	}
+	if _, ok, err := ReviewedHiddenAliasTarget(cmd, "alias"); ok || err != nil {
+		t.Fatalf("unannotated hidden flag = ok %v, err %v", ok, err)
+	}
+
+	for _, tc := range []struct {
+		name   string
+		target string
+		origin string
+		setup  func(*cobra.Command)
+		want   string
+	}{
+		{name: "unreviewed origin", target: "public", origin: "other"},
+		{name: "missing target", origin: runtimeannotate.FlagAliasOriginCorecmdV1, want: "malformed reviewed alias target"},
+		{name: "self", target: "alias", origin: runtimeannotate.FlagAliasOriginCorecmdV1, want: "cannot alias itself"},
+		{name: "unknown target", target: "missing", origin: runtimeannotate.FlagAliasOriginCorecmdV1, want: "unknown executable input"},
+		{name: "hidden target", target: "hidden", origin: runtimeannotate.FlagAliasOriginCorecmdV1, setup: func(cmd *cobra.Command) {
+			cmd.Flags().String("hidden", "", "hidden")
+			_ = cmd.Flags().MarkHidden("hidden")
+		}, want: "targets hidden input"},
+		{name: "alias target", target: "public", origin: runtimeannotate.FlagAliasOriginCorecmdV1, setup: func(cmd *cobra.Command) {
+			runtimeannotate.SetFlagAnnotation(cmd.Flags().Lookup("public"), runtimeannotate.AnnotationFlagAliasOf, "other")
+		}, want: "targets alias input"},
+		{name: "type mismatch", target: "number", origin: runtimeannotate.FlagAliasOriginCorecmdV1, setup: func(cmd *cobra.Command) {
+			cmd.Flags().Int("number", 0, "number")
+		}, want: "incompatible types"},
+	} {
+		t.Run(tc.name, func(t *testing.T) {
+			cmd := newCommand()
+			if tc.setup != nil {
+				tc.setup(cmd)
+			}
+			annotate(cmd, tc.target, tc.origin)
+			_, ok, err := ReviewedHiddenAliasTarget(cmd, "alias")
+			if tc.want == "" {
+				if ok || err != nil {
+					t.Fatalf("ok = %v, err = %v", ok, err)
+				}
+				return
+			}
+			if ok || err == nil || !strings.Contains(err.Error(), tc.want) {
+				t.Fatalf("ok = %v, err = %v, want %q", ok, err, tc.want)
+			}
+		})
+	}
+
+	cmd = newCommand()
+	annotate(cmd, "public", runtimeannotate.FlagAliasOriginCorecmdV1)
+	if target, ok, err := ReviewedHiddenAliasTarget(cmd, "alias"); target != "public" || !ok || err != nil {
+		t.Fatalf("reviewed alias = %q, %v, %v", target, ok, err)
+	}
+}
